Write a function to convert string to u32
Define a function named `str_to_u32` in the 'main.c'. The function takes
a null terminated string as a parameter and converts it to an unsigned
32 bit integer. In case of invalid format, it returns U32_MAX to
indicate error.

Using U32_MAX as error code wouldn't be a problem because our program
doesn't need to deal with such a large number.

Instead of using `scanf()` to take the integer from the input, take the
whole line from input and convert it to u32 using this function. This
makes sure that user can only input one number in one line.