Write the basic `getchar()` function
Write function `getchar()` that takes a buffer of size 121 and reads a
line from the console into the buffer. The max characters read are 120
as the last char is reserved for the null character.

Use this function for taking the number input from the console and
convert them into u32 using `str_to_u32()` function.