Write `prompt()` function to ask for input
Define a function called `prompt()` that takes a const string as input,
prints it on the console, and waits for user to enter a valid u32
number. It converts the u32 number string to an u32, and returns it.

We can now use this `prompt()` function to simply take user input in a
single statement.