Handle inccorect param values and bad inputs
Move the common inccorect format error to the `prompt()` function. Check
for the inccorect values of the parameters and give appropriate error
messages.

This will make sure that the player can only put valid pamaters and also
refactors the code by putting the handling of number format in the
`prompt()` function.